requirements.txt
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
idna==2.10
Pillow==8.0.1
pytesseract==0.3.6
requests==2.25.0
urllib3==1.26.2
